pyproject.toml
[virtualenvs]
in-project = true

[build-system]
requires = [ "poetry-core>=1.0.0",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
version = "0.1.0"
name = "airbyte-source-YOUR_PACKAGE"
description = "Source implementation for YOUR_PACKAGE."
authors = [ "YOUR_GITHUB_USERNAME",]
license = "MIT"
readme = "README.md"
repository = "https://github.com/YOUR_GITHUB_USERNAME/YOUR_PACKAGE"
packages = [ { include = "airbyte_source_YOUR_PACKAGE" }, {include = "main.py" } ]

[tool.poetry.dependencies]
python = "^3.9,<3.12"
airbyte-cdk = "^0"

[tool.poetry.scripts]
airbyte-source-YOUR_PACKAGE = "airbyte_source_YOUR_PACKAGE.run:run"

[tool.poetry.group.dev.dependencies]
requests-mock = "*"
pytest-mock = "*"
pytest = "*"
black = "^24.10.0"
isort = "^5.13.2"

